feat: Initial implementation of base85
This implementation provides a Base85 implementation with an API similar to encoding/base64 so that custom alphabets and paddings can be provided.
Using this we provide RFC1924 as a standard alphabet.

This is being provided because the standard `ascii85` implementation does not allow custom character sets that would enable use of RFC1924 or other denser encodings (compared to Base64 which is the most dense built-in encoding) provided in standard go.